Move instance repository definition to repository.go
