+++ broadway/instance.go
@@ -1,10 +1,6 @@
 package broadway
 
-import (
-	"encoding/json"
-
-	"github.com/namely/broadway/store"
-)
+import "encoding/json"
 
 // Instance entity
 type Instance struct {
@@ -26,32 +22,4 @@ func (i *Instance) JSON() (string, error) {
 // Path for an instance
 func (i *Instance) Path() string {
 	return "/broadway/instances/" + i.PlaybookID + "/" + i.ID
-}
-
-// Instance repository interface
-type InstanceRepository interface {
-	Save(instance Instance) error
-}
-
-// Instance repository to handle persistence logic
-type InstanceRepo struct {
-	store store.Store
-}
-
-// Create a new instance repo
-func NewInstanceRepo(s store.Store) *InstanceRepo {
-	return &InstanceRepo{store: s}
-}
-
-// Save a new instance as json
-func (ir *InstanceRepo) Save(instance Instance) error {
-	encoded, err := instance.JSON()
-	if err != nil {
-		return err
-	}
-	err = ir.store.SetValue(instance.Path(), encoded)
-	if err != nil {
-		return err
-	}
-	return nil
 }
